% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createterms.R
\name{generate_conditions}
\alias{generate_conditions}
\title{Generate Search Conditions}
\usage{
generate_conditions(and_conditions, or_group = NULL)
}
\arguments{
\item{and_conditions}{Ein Vektor mit AND-Bedingungen, jede AND-Bedingung enthält OR-Gruppen.}

\item{or_group}{(Optional) Eine spezifische OR-Gruppe, die gesucht werden soll.}
}
\value{
Eine Liste von Bedingungen, die in apply_term_search verwendet werden können.
}
\description{
Diese Funktion erzeugt die Suchbedingungen basierend auf den Benutzereingaben.
Der Benutzer kann angeben, ob er die gesamte AND-Bedingung oder nur eine OR-Gruppe innerhalb einer AND-Bedingung verwenden möchte.
}
