% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/more_or.R
\name{moreor_apply_search_with_context}
\alias{moreor_apply_search_with_context}
\title{Search multiple OR-Groups Across Multiple Documents with Context (Parallel)}
\usage{
moreor_apply_search_with_context(corpus, conditions)
}
\arguments{
\item{corpus}{Ein Korpus von Dokumenten (Liste von Zeichenketten).}

\item{conditions}{Eine Liste von Bedingungen.}
}
\value{
Eine Liste der relevanten Dokumente mit den gefundenen Begriffen und Kontext.
}
\description{
Diese Funktion durchsucht mehrere Dokumente parallel nach Begriffen und berücksichtigt dabei den Kontext (eine Seite davor und danach).
}
