% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_or.R
\name{oneor_process_findings}
\alias{oneor_process_findings}
\title{Process Findings of a single OR-Group}
\usage{
oneor_process_findings(relevant_documents)
}
\arguments{
\item{relevant_documents}{Eine Liste von relevanten Dokumenten (wie von oneor_apply_search ausgegeben).}
}
\value{
Ein DataFrame mit den formatierten Ergebnissen.
}
\description{
Diese Funktion verarbeitet die Ergebnisse der Suche mit einer OR-Gruppe und formatiert sie in einen
Datensatz, der Mehrfach-Treffer pro Dokument beruecksichtigt und in ein breites Format bringt.
}
