% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleandata.R
\name{process_findings}
\alias{process_findings}
\title{Process Findings and Transform into Dataset}
\usage{
process_findings(relevant_documents)
}
\arguments{
\item{relevant_documents}{Eine Liste von relevanten Dokumenten (wie von apply_term_search zurückgegeben).}
}
\value{
Ein DataFrame mit den formatierten Ergebnissen.
}
\description{
Diese Funktion verarbeitet die Suchergebnisse (relevant_documents) und formatiert sie in einen
Datensatz, der Mehrfach-Treffer pro Dokument berücksichtigt und in ein breites Format bringt.
}
